Fix log formatter context usage.
The log formatter context must be called when used in a `with` statement.

Fixes ZEN-33032.
