+++ Products/Jobber/tests/test_log.py
@@ -24,6 +24,7 @@ from ..log import (
     load_log_level_config,
     _loglevelconf_filepath,
 )
+from .utils import LoggingLayer
 
 UNEXPECTED = type("UNEXPECTED", (object,), {})()
 PATH = {"src": "Products.Jobber.log"}
@@ -182,22 +183,6 @@ class LoadLogLevelConfigTest(TestCase):
         actual = load_log_level_config(filename)
 
         t.assertDictEqual(expected, actual)
-
-
-class LoggingLayer(object):
-    """Test layer to support testing with Python's logging API."""
-
-    @classmethod
-    def setUp(cls):
-        cls.original_manager = logging.Logger.manager
-        cls.manager = logging.Manager(logging.root)
-        logging.Logger.manager = cls.manager
-
-    @classmethod
-    def tearDown(cls):
-        logging.Logger.manager = cls.original_manager
-        del cls.manager
-        del cls.original_manager
 
 
 class ApplyLevelsTest(TestCase):
